refactor calcutor logic into service
